Add meanings; show example keyword(s) in bold
This code loops through the Dictionary API response a few times to display various definitions with their example sentences (if applicable). Definitions are structured using semantic HTML elements and numerical indices. Example sentences are styled to appear distinct from the definitions. I experimented successfully with getting the keyword to appear in boldface for almost all cases. The notable exception is where the example uses the keyword in a different tense that drastically alters the spelling (e.g. "swung" will not appear bold if the keyword was "swing," although "appropriated" will work if the keyword was "appropriate").