Add Pexels API & log response; update form
Improvement to UI: The form will be reset upon submission so that the search word will not persist in the input field across renderings. That eliminates the confusing experience of seeing the previous search term when, for example, a new word is searched for by double-clicking a synonym or antonym.